+++ src/Form/LotteryVideoType.php
@@ -1,0 +1,53 @@
+<?php
+
+/*
+ * (c) 2025: 975L <contact@975l.com>
+ * (c) 2025: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ShopBundle\Form;
+
+use Symfony\Component\Form\AbstractType;
+use c975L\ShopBundle\Entity\LotteryVideo;
+use Symfony\Component\Form\FormBuilderInterface;
+use Vich\UploaderBundle\Form\Type\VichFileType;
+use Symfony\Component\Validator\Constraints\File;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class LotteryVideoType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('file', VichFileType::class, [
+                'label' => 'Video',
+                'required' => false,
+                'allow_delete' => true,
+                'download_uri' => true,
+                'asset_helper' => true,
+                'constraints' => [
+                    new File([
+                        'maxSize' => '500M',
+                        'mimeTypes' => [
+                            'video/mp4',
+                            'video/webm',
+                            'video/ogg',
+                        ],
+                        'mimeTypesMessage' => 'Please upload a valid video file',
+                    ])
+                ],
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => LotteryVideo::class,
+            'translation_domain' => 'shop'
+        ]);
+    }
+}